requirements.txt
tiktoken
faiss-cpu
Flask==2.3.2
requests==2.31.0
python-dotenv==1.0.0
openai
langchain
numpy==1.24.3
PyDrive==1.3.1
pinecone-client==2.2.4
Flask-Cors==4.0.0
wheel==0.41.2
pypdf==3.14.0
twilio==8.10.2
Flask-PyMongo==2.3.0
